requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
Django==4.2
django-crispy-forms==2.0
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.1.1
pytz==2023.3
requests==2.29.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
